+++ src/Oro/Bundle/PromotionBundle/Tests/Functional/Model/CouponApplicabilityQueryBuilderModifierTest.php
@@ -32,7 +32,7 @@ class CouponApplicabilityQueryBuilderModifierTest extends WebTestCase
 
         $queryBuilder = $repository->createQueryBuilder('c')->orderBy('c.usesPerCoupon');
         $coupons = $queryBuilder->getQuery()->getResult();
-        $this->assertCount(6, $coupons);
+        $this->assertCount(7, $coupons);
 
         $this->modifier->modify($queryBuilder);
         $coupons = $queryBuilder->getQuery()->getResult();
@@ -41,7 +41,7 @@ class CouponApplicabilityQueryBuilderModifierTest extends WebTestCase
             [
                 $this->getReference(LoadCouponData::COUPON_WITHOUT_PROMO_AND_VALID_UNTIL),
                 $this->getReference(LoadCouponData::COUPON_WITH_PROMO_AND_WITHOUT_VALID_UNTIL),
-                $this->getReference(LoadCouponData::COUPON_WITH_PROMO_AND_VALID_UNTIL),
+                $this->getReference(LoadCouponData::COUPON_WITH_PROMO_AND_VALID_FROM_AND_UNTIL),
                 $this->getReference(LoadCouponData::COUPON_WITH_SHIPPING_PROMO_AND_VALID_UNTIL)
             ],
             $coupons